Update patch set 1
Patch Set 1: Do not submit

Yep, you are right, better -2 :)

Well you are also correct with the blobs and the permission, but i thought for development it's nicer to have all in one change. 
When it works in a good way i will ask for permission @ Sony and then split it to correct changes.

Patch-set: 1