Update patch set 1
Patch Set 1: Code-Review-1

I would prefer to have the audio_policy change in another file, and to not override CM changes.


Even if stock firmware contains other files, only snd_soc_msm_2x_Fusion3 would be used for our platform (not because it's sony fusion3, but as it's qcom fusion3 too).


Which build are you experiencing that with?
The fusion3-common repo got in a bad state (cm-11.0-new branch only) earlier due to a botched rebase, but that has been fixed now.

Patch-set: 1
Label: Code-Review=-1